+++ Changelog.md
@@ -3,9 +3,9 @@
 ###In progress
 
 ###5.35
-Add steps for comments and collapsible bodies
-Log debug message when comparison fails due to unintentional special regex characters instead of warning
-Check if debug is enabled on logger before checking for regex
+Add steps for comments and collapsible bodies  
+Log debug message when comparison fails due to unintentional special regex characters instead of warning  
+Check if debug is enabled on logger before checking for regex  
 
 ###5.34
 Add warning message when JSON comparison fails due to unintentional special regex characters